requirements.txt
telethon
flask
psycopg2-binary==2.9.9